Modules/GUI/requirements.txt
dearpygui
rplidar-roboticia
requests
fastapi
uvicorn
numpy
